Conflicts:
	433Net/433Net/433Net/player.cpp

+++ 433Net/433Net/433Net/usertoken.h
@@ -16,9 +16,8 @@ public:
 
 	char buf[BUFSIZE];
 	int position;
-	short remainBytes;
+	int remainBytes;
 	bool var;
-
 	IPeer* peer;
 
 public:
